+++ test/test_note.py
@@ -49,7 +49,7 @@ def test_create_note(barerepo):
     annotated_id = barerepo.revparse_single('HEAD~3').id
     author = committer = Signature('Foo bar', 'foo@bar.com', 12346, 0)
     note_id = barerepo.create_note(NOTE[1], author, committer, str(annotated_id))
-    assert NOTE[0] == str(note_id)
+    assert NOTE[0] == note_id
 
     # check the note blob
     assert NOTE[1].encode() == barerepo[note_id].data
@@ -58,7 +58,7 @@ def test_create_note(barerepo):
 def test_lookup_note(barerepo):
     annotated_id = str(barerepo.head.target)
     note = barerepo.lookup_note(annotated_id)
-    assert NOTES[0][0] == str(note.id)
+    assert NOTES[0][0] == note.id
     assert NOTES[0][1] == note.message
 
 
@@ -73,7 +73,7 @@ def test_remove_note(barerepo):
 
 def test_iterate_notes(barerepo):
     for i, note in enumerate(barerepo.notes()):
-        assert NOTES[i] == (str(note.id), note.message, str(note.annotated_id))
+        assert NOTES[i] == (note.id, note.message, note.annotated_id)
 
 
 def test_iterate_non_existing_ref(barerepo):