+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - New `push_options` optional argument in `Repository.push(...)`
   [#1282](https://github.com/libgit2/pygit2/pull/1282)
 
+- New support comparison of `Oid` with text string
+
 - Remove setuptools runtime dependency
   [#1281](https://github.com/libgit2/pygit2/pull/1281)